pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cp-tox-mini"
version = "0.1.0"
description = "Cell Painting × toxicity fusion workflow"
authors = [{name = "Bio-ML Engineer", email = "engineer@example.com"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"

dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "joblib>=1.0.0",
    "pillow>=8.0.0",
    "pyarrow>=5.0.0",
    "typer>=0.4.0",
    "pydantic>=1.8.0",
    "markdown>=3.3.0",
    "tqdm>=4.60.0",
    "rdkit-pypi>=2022.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
    "flake8>=3.9.0",
    "black>=21.0.0",
    "mypy>=0.910",
]

[project.scripts]
cp-tox-mini = "cp_tox_mini.cli:main"

[tool.setuptools.packages.find]
include = ["cp_tox_mini*"]

[tool.setuptools.package-data]
cp_tox_mini = ["*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true